+++ CHANGELOG.md
@@ -38,6 +38,11 @@ Known Limitations
   - Peer-to-Peer Rooms
 - When the Room is being recorded, the SFU will not disable any simulcast layers of the publisher's VideoTrack.
 
+Bug Fixes
+---------
+
+- Fixed a bug where setting clientTrackSwitchOffControl to `auto` caused the RemoteVideoTrack's to get switched off even while playing in picture-in-picture window. (VIDEO-6677)
+  Note that this fix does not work on firefox because firefox does not yet implement [picture-in-picture](https://developer.mozilla.org/en-US/docs/Web/API/Picture-in-Picture_API) APIs.
 
 2.18.1 (October 29, 2021)
 =========================